+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- A raw collections used as parameter in an implementation was not recognised as collection
+  (see GitLab [issue 6](https://gitlab.ilabt.imec.be/fno/proc/function-component/-/issues/6)). 
+
 ## [0.0.3] - 2022-05-13
 
 ### Changed